refactor(controllers): move controllers to controllers folder
