+++ src/services/healthcheck.db.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from './infrastructure/orm/prisma.service';
+import { PrismaService } from '../infrastructure/orm/prisma.service';
 
 @Injectable()
 export class HealthCheckDBService {
