+++ src/types/FormQuestion.ts
@@ -5,14 +5,10 @@ interface ConditionalTypeInfo {
   value_range?: number;
 }
 
-interface FormQuestion {
+export interface FormQuestion {
   id: string;
   type: string;
   heading: string;
   content?: string;
   conditional?: ConditionalTypeInfo;
-}
-
-export default interface RoutineForm {
-  questions: FormQuestion[];
 }
